requirements.txt
Flask
flask-jsonpify
flask-restful
python-jenkins
pyyaml
pytest
